feat: implement upsertUsername
