+++ .gitignore
@@ -18,3 +18,4 @@
 
 yarn-debug.log*
 yarn-error.log*
+todo.md
