+++ b4f/src/db/index.ts
@@ -11,7 +11,7 @@ export const dbFactory = (dbType: DbType): Connector => {
 export interface Connector {
   getUserInfo: (userId: string) => Promise<UserInfoDTO>
   checkUsername: (username: string) => Promise<boolean>
-  setUsername: (username: string, userId: string) => Promise<void>
+  upsetUsername: (username: string, userId: string) => Promise<void>
   getCampaigns: (userId: string) => Promise<GetCampaignsDTO>
   createCampaign: (userId: string, name: string) => Promise<CampaignDTO>
   editCampaign: (id: string, userId: string, name: string, status: CampaignStatus) => Promise<void>
