+++ b4f/src/db/postgreDb/user.ts
@@ -44,7 +44,7 @@ export const checkUsername = async (username: string): Promise<boolean> => {
       WHERE username = $1
     ) as exists;
   `
-  const checkUsernameValues = [username]
+  const checkUsernameValues = [username.toLocaleLowerCase()]
   const check = await client.query<DBCheckUsername>(checkUsernameQuery, checkUsernameValues)
   client.release()
 
@@ -53,14 +53,16 @@ export const checkUsername = async (username: string): Promise<boolean> => {
   return !check.rows[0].exists
 }
 
-export const setUsername = async (username: string, google_id: string): Promise<void> => {
+export const upsetUsername = async (username: string, google_id: string): Promise<void> => {
   const client = await dbConfig.connect()
-  const inserUsernameQuery = `
-    INSERT INTO ${tableUsers}
-    (username, google_id)
-    VALUES ($1, $2);
-  `
-  const insertUsernameValues = [username, google_id]
-  await client.query(inserUsernameQuery, insertUsernameValues)
+  const upsertUsernameQuery = `
+  INSERT INTO ${tableUsers}
+  (username, google_id)
+  VALUES ($1, $2)
+  ON CONFLICT (google_id) DO UPDATE
+  SET username = EXCLUDED.username;
+`
+  const upsertUsernameValues = [username.toLocaleLowerCase(), google_id]
+  await client.query(upsertUsernameQuery, upsertUsernameValues)
   client.release()
 }
