+++ docker-compose.yml
@@ -1,4 +1,3 @@
-version: "3.8"
 volumes:
   rpg_pal_data:
 services:
