+++ frontend/src/Controllers/Home.tsx
@@ -18,7 +18,7 @@ import { clearMasterState } from '../Store/master'
 import { parseErrorMessage } from '../Utils/f'
 import { useNavigate } from 'react-router-dom'
 import useGoogleLoginWithRedux from '../Hooks/useGoogleLoginWithRedux'
-import InsertUsername from './InsertUsername'
+import InsertUsername from '../Components/InsertUsername'
 import Loader from '../Components/Loader'
 import HomeInfo from '../Components/HomeInfo'
 import ErrorComponent from '../Components/Error'
